Change project structure to cmd/pkg and refactor code accordingly
