+++ pkg/taon/columns_value_test.go
@@ -1,4 +1,4 @@
-package main
+package taon
 
 import (
 	"reflect"
